Refactor CustomTextField, proptype description in readme.me file
