+++ frontend/src/components/TextFieldComponents/CustomTextField/CustomTextField.jsx
@@ -1,7 +1,8 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { TextField, Chip, Box, InputLabel } from "@mui/material";
-import "./TextFieldStyles.css";
+import ChipAdornment from "../Chips/ChipAdornment";
+import "./CustomTextFieldStyles.css";
 
 const CustomTextField = ({
   labelText,
@@ -20,9 +21,7 @@ const CustomTextField = ({
 }) => {
   return (
     <Box paddingY={1}>
-      <InputLabel className="label" sx={{ fontWeight: labelFontWeight }}>
-        {labelText}
-      </InputLabel>
+      <InputLabel sx={{ fontWeight: labelFontWeight }}>{labelText}</InputLabel>
       <TextField
         className="textField"
         sx={{ width: TextFieldWidth }}
@@ -39,26 +38,7 @@ const CustomTextField = ({
           endAdornment: endAdornment,
           ...(chips &&
             chips.length > 0 && {
-              startAdornment: (
-                <Box
-                  sx={{
-                    display: "flex",
-                    gap: 1,
-                    mt: -15,
-                    mr: 1,
-                  }}
-                >
-                  {chips.map((chip, index) => (
-                    <Chip
-                      key={index}
-                      label={chip.label}
-                      onDelete={chip.onDelete}
-                      variant="outlined"
-                      sx={{ borderRadius: "5px" }}
-                    />
-                  ))}
-                </Box>
-              ),
+              startAdornment: <ChipAdornment chips={chips} />,
             }),
         }}
         inputProps={{
@@ -75,23 +55,6 @@ const CustomTextField = ({
     </Box>
   );
 };
-
-// Explanation of the PropTypes used in this file
-/*
-- labelText (string): A label that describes the content of the text field.
-- defaultValue (string): A default value for when the input is still empty.
-- helperText (string): The text that will be displayed as the helper text for the text field.
-- error (bool): A boolean value that determines whether the text field should display an error state.
-- multiline (bool): A boolean value that determines whether the text field should allow multiple lines of text.
-- rows (number): The number of rows that the text field should display when in multiline mode.
-- startAdornment (node): The node that will be displayed as the start adornment for the text field.
-- endAdornment (node): The node that will be displayed as the end adornment for the text field.
-- placeholder (string): The short hint displayed in the input before the user enters a value.
-- chips (array): An array of objects that represent chips to be displayed in the text field.
-- labelFontWeight (number): The font weight of the label text.
-- TextFieldWidth (string): The width of the text field.
-- inputHeight (string): The height of the input field.
-*/
 
 CustomTextField.defaultProps = {
   labelText: "",
